third_party/blink/web_tests/fast/css-intrinsic-dimensions/border-box-sizing.html

<!DOCTYPE html>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<style>
.container {
  box-sizing: border-box;
}
</style>

<body onload="checkLayout('.container')">
<div id=log></div>

<div style="display: flex; flex-direction: column; height: 20px;">
  <!-- uses the implied min-height: auto; should not flex-shrink below 80 -->
  <div class="container" style="padding-top: 30px;" data-expected-height="80">
     <div style="height: 50px; width: 50px;"></div>
  </div>
</div>

<div class="container" style="position: absolute; height: 20px; width: min-content; padding-left: 30px;" data-expected-width="80">
   <div style="height: 50px; width: 50px;"></div>
</div>


<div class="container" style="height: 20px; width: min-content; padding-left: 30px;" data-expected-width="80">
   <div style="height: 50px; width: 50px;"></div>
</div>
